Backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<11.0
gunicorn>=21.2,<22.0
dj-database-url>=2.1,<3.0
